% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\name{ATACciceroPlot}
\alias{ATACciceroPlot}
\title{Regulatory Potential Plot}
\usage{
ATACciceroPlot(
  mtx,
  score,
  pval,
  selected = NULL,
  DA_peaks = NULL,
  species = "Hs",
  version = "v75",
  gene_annotations = NULL,
  pro_up = 2000,
  pro_down = 0,
  overlap_cut = 0,
  NTC = "NTC",
  min.pct = 0.1,
  test.use = "wilcox",
  p_adj_cut = 0.05,
  logFC_cut = 0.25,
  score_cut = 0,
  pval_cut = 0.05,
  upstream = 2e+06,
  downstream = 2e+06,
  track_size = c(1, 0.3, 0.2, 0.3),
  include_axis_track = TRUE,
  connection_color = "#7F7CAF",
  connection_color_legend = TRUE,
  connection_width = 2,
  connection_ymax = NULL,
  gene_model_color = "#81D2C7",
  alpha_by_coaccess = FALSE,
  gene_model_shape = c("smallArrow", "box"),
  plot.save = TRUE,
  table.save = TRUE,
  prefix = ".",
  label = "",
  width = 7,
  height = 7,
  png_res = 720,
  html_config = FALSE
)
}
\arguments{
\item{mtx}{SeuratObject or directory to rds file of SeuratObject, with cell in columns and features in rows.}

\item{score}{Data frame or directory of score from \code{improved_scmageck_lr}.}

\item{pval}{Data frame or directory of p_value from \code{improved_scmageck_lr}.}

\item{selected}{Perturbation to find DA peaks. Default is \code{NULL}, all perturbations will be chosen.}

\item{species}{Only support "Hs" and "Mm". Default is "Hs".}

\item{version}{Version of the reference genome(Ensembl). Default is "v75".}

\item{gene_annotations}{Gene annotations stored in data frame format, including c("chromosome", "start", "end", "strand", "transcript") as colnames. Default is \code{NULL}, gene annotations are from \code{\link{ensembldb}}.}

\item{pro_up}{The number of nucleotides upstream of the transcription start site that should be included in the promoter region. Default is 3000.}

\item{pro_down}{The number of nucleotides downstream of the transcription start site that should be included in the promoter region. Default is 0.}

\item{overlap_cut}{Maximum overlap nucleotides between peaks and promoters. Default is 0.}

\item{NTC}{The name of negative controls. Default is "NTC".}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the selected perturbation or NTC. Meant to speed up the function by not testing genes that are very infrequently expressed. Default is 0.1.}

\item{test.use}{Denotes which test to use. Available options are: "wilcox", "bimod", "roc", "t", "negbinom", "poisson", "LR", "MAST", "DESeq2". See details from \code{\link[Seurat]{FindMarkers}}. Default is "wilcox".}

\item{p_adj_cut}{Maximum adjust p_value. Default is 0.05.}

\item{logFC_cut}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.25. Increasing logfc.threshold speeds up the function, but can miss weaker.}

\item{score_cut}{Score cutoff of \code{improved_scmageck_lr} results. Default is 0.}

\item{pval_cut}{P-value cutoff of \code{improved_scmageck_lr} results. Default is 0.05.}

\item{upstream}{The number of nucleotides upstream of the start site of selected region. Default is 2000000.}

\item{downstream}{The number of nucleotides downstream of the start site of selected region. Default is 2000000.}

\item{track_size}{Size of each axis. Default is c(1,.3,.2,.3). If \code{include_axis_track=FALSE}, track_size should be a vector with 3 elements.}

\item{include_axis_track}{Logical, should a genomic axis be plotted? Default is \code{TRUE}.}

\item{connection_color}{Color for connection lines. A single color, the name of a column containing color values, or the name of a column containing a character or factor to base connection colors on. Default is "#7F7CAF".}

\item{connection_color_legend}{Logical, should connection color legend be shown? Default is \code{TRUE}.}

\item{connection_width}{Width of connection lines. Default is 2}

\item{connection_ymax}{Connection y-axis height. Default is NULL, chosen automatically.}

\item{gene_model_color}{Color for gene annotations. Default is "#81D2C7".}

\item{alpha_by_coaccess}{Logical, should the transparency of connection lines be scaled based on co-accessibility score? Default is \code{FALSE}.}

\item{plot.save}{Logical, save plots or not. Default is \code{TRUE}.}

\item{table.save}{Logical, save DA peaks table or not. Default is \code{TRUE}.}

\item{prefix}{Path to save the plots. Default is current directory.}

\item{label}{The prefix label of the output file. Notably, there needs a separator between default file names and the label, so label would be better to be like "label_". Default is "".}

\item{width}{Width of the graphics region of the pdf file in inches, for both png and pdf format. Default is 7.}

\item{height}{Height of the graphics region of the pdf file in inches, for both png and pdf format. Default is 7.}

\item{png_res}{The nominal resolution in ppi of png file. Higher png_res indicates a bigger and more clear png file. Default is 720.}

\item{html_config}{Logical, generate and return a list includes the config character string of html. Default is \code{FALSE}.}

\item{DApeaks}{Data frame or directory of DA peaks table generated before. Need a additional column of perturbation information, named "TF". Default is \code{NULL}, will find DA peaks for selected perturbations.}

\item{gene_model_shap}{Character scalar. The shape in which to display the track items. Currently only box, arrow, ellipse, and smallArrow are implemented. Default is c("smallArrow", "box").}
}
\description{
Visualize the relationship between potential enhancers and genes, using results from \code{\link[SCREE]{improved_scmageck_lr}}. Modifies from \code{\link[cicero]{plot_connections}}. For scATAC-seq based input, potential enhancers are defined as DA peaks without overlap with promoter regions.
}
