% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAP.R
\name{ATACumap}
\alias{ATACumap}
\title{Run UMAP and Visualization}
\usage{
ATACumap(
  mtx,
  assays = "peak",
  min.cutoff = "q5",
  reduction.key = "LSI_",
  reduction.name = "lsi",
  nsvs = 50,
  n = 20,
  dims = NULL,
  dims.cut = 0.5,
  algorithm = 1,
  resolution = 0.8,
  title.size = 25,
  legend.key.size = unit(0.7, "cm"),
  legend.text.size = 14,
  x.text.size = 16,
  x.title.size = 20,
  y.text.size = 16,
  y.title.size = 20,
  pt.size = 0.2,
  raster = FALSE,
  label.cut = 20,
  plot.show = TRUE,
  plot.return = FALSE,
  plot.save = TRUE,
  prefix = ".",
  label = "",
  width = 3.5,
  height = 3.5,
  png_res = 720
)
}
\arguments{
\item{mtx}{SeuratObject or directory to rds file of SeuratObject, with cell in columns and features in rows. Note that the dataset has to be normalized and scaled, and need a sgRNA information column named "perturbations" in the metadata.}

\item{assays}{Assay to use. Default is "peak".}

\item{min.cutoff}{Cutoff for feature to be included in the VariableFeatures for the object. This can be a percentile specified as 'q' followed by the minimum percentile, for example 'q5' to set the top 95\% most common features as the VariableFeatures for the object. Alternatively, this can be an integer specifying the minimum number of cells containing the feature for the feature to be included in the set of VariableFeatures. For example, setting to 10 will include features in >10 cells in the set of VariableFeatures. If NULL, include all features in VariableFeatures. If NA, VariableFeatures will not be altered, and only the feature metadata will be updated with the total counts and percentile rank for each feature.}

\item{reduction.key}{Key for dimension reduction object. Default is "LSI_".}

\item{reduction.name}{Name for stored dimension reduction object. Default is "lsi".}

\item{nsvs}{Number of singular values to compute. Default is 50.}

\item{n}{Number of components to use. Default is 20. If NULL, use all components.}

\item{dims}{Dimensions of reduction to use as input. Default is \code{NULL}, which means use all components with a correlation < \code{dims.cut} between total counts, from \code{n} components used.}

\item{algorithm}{Algorithm for modularity optimization (1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm). Leiden requires the leidenalg python. Default is 1.}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities. Default is 0.8.}

\item{title.size}{Numeric, title size of the UMAP. Default is 25.}

\item{legend.key.size}{Numeric, legend key size of the UMAP. Default is unit(0.7, "cm").}

\item{legend.text.size}{Numeric, legend text size of the UMAP. Default is 14.}

\item{x.text.size}{Numeric, x-axis text size of the UMAP. Default is 16.}

\item{x.title.size}{Numeric, x-axis title size of the UMAP. Default is 20.}

\item{y.text.size}{Numeric, y-axis text size of the UMAP. Default is 16.}

\item{y.title.size}{Numeric, y-axis title size of the UMAP. Default is 20.}

\item{pt.size}{Point size of the UMAP. Default is 0.2.}

\item{raster}{Logical, convert points to raster format, will be useful to reduce the storage cost of the output figure or pdf. Default is \code{FALSE}.}

\item{label.cut}{Maximum perturbations to be labeled in the UMAP. Default is 20, which means to label top 20 perturbations ranked by cell numbers.}

\item{plot.show}{Logical, show the UMAP plot or not. Default is \code{TRUE}.}

\item{plot.return}{Logical, return a list include the SeuratObject after running UMAP and two UMAP plots or not. Default is \code{FALSE}.}

\item{plot.save}{Logical, save plots or not. Default is \code{TRUE}.}

\item{prefix}{Path to save the plots. Default is current directory.}

\item{label}{The prefix label of the output file. Notably, there needs a separator between default file names and the label, so label would be better to be like "label_". Default is "".}

\item{width}{Width of the graphics region of the pdf file in inches, for both png and pdf format. Default is 3.5.}

\item{height}{Height of the graphics region of the pdf file in inches, for both png and pdf format. Default is 3.5.}

\item{png_res}{The nominal resolution in ppi of png file. Higher png_res indicates a bigger and more clear png file. Default is 720.}
}
\description{
Run UMAP from SeuratObject after normalization and scale, and perform visualization for scATAC-seq based input, based on \code{\link{Seurat}}.
}
