% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAC.R
\name{CalculateGeneActivity}
\alias{CalculateGeneActivity}
\title{Calculate Gene Activity}
\usage{
CalculateGeneActivity(
  mtx,
  fragments,
  species = "Hs",
  version = "v75",
  gene_type = "Symbol",
  protein_coding = TRUE,
  pro_up = 3000,
  pro_down = 0,
  sep = c("-", "-")
)
}
\arguments{
\item{mtx}{SeuratObject or directory to rds file of SeuratObject, with cell in columns and features in rows.}

\item{fragments}{Directory of fragments file or a FragmentObject generated from \code{\link[Signac]{CreateFragmentObject}}.}

\item{species}{Only support "Hs" and "Mm". Default is "Hs".}

\item{version}{Version of the reference genome(Ensembl). Default is "v75".}

\item{gene_type}{Type of gene name, selected from one of c("Symbol", "Ensembl"). Default is "Symbol".}

\item{protein_coding}{Logical, only use protein coding gene or not. Default is \code{TRUE}.}

\item{pro_up}{Numeric, the number of nucleotides upstream of the transcription start site that should be included in the promoter region. Default is 3000.}

\item{pro_down}{Numeric, the number of nucleotides downstream of the transcription start site that should be included in the promoter region. Default is 0.}

\item{sep}{Vector of separators to use for genomic string. First element is used to separate chromosome and coordinates, second separator is used to separate start and end coordinates.}
}
\description{
Generate gene activity matrix using peak count matrix.
}
