% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb_enrichment.R
\name{CalculatePerturbEnrichment}
\alias{CalculatePerturbEnrichment}
\title{Calculate Perturbation Enrichment}
\usage{
CalculatePerturbEnrichment(
  mtx,
  sg_lib,
  NTC = "NTC",
  NTC.cal = TRUE,
  table.save = TRUE,
  top = NULL,
  range = c(0, 1),
  color = c("white", "coral1"),
  cell = c(15, 20),
  fontsize = 15,
  angle = 90,
  legend.title = FALSE,
  plot.save = TRUE,
  prefix = ".",
  label = "",
  width = 10,
  height = 8,
  png_res = 720
)
}
\arguments{
\item{mtx}{SeuratObject or directory to rds file of SeuratObject, with cell in columns and features in rows. Note that the dataset has to be normalized and scaled, and need a sgRNA information column named "perturbations" in the meta data.}

\item{sg_lib}{Data frame or directory to a txt file containing 3 columns: cell, barcode, gene. If sgRNA information stored in a matrix-like format or input data frame only has sgRNA frequency of each cell, use \code{\link[SCREE]{sgRNAassign}} to assign sgRNA to each cell.}

\item{NTC}{The name of negative controls. Default is "NTC".}

\item{NTC.cal}{Logical, calculate negative control or not. Default is \code{TRUE}, which will remove cells assigned with negative control from the SeuratObject.}

\item{table.save}{Logical, save perturbation enrichment table or not. Default is \code{TRUE}.}

\item{top}{Top perturbations with the highest enrichment ratio to visualize. Default is \code{NULL}, using all perturbations to visualize.}

\item{range}{Range of the ratio. Since for some datasets, the max enrichment ratio among all perturbations is extremely small, the range can be set to a smaller limitation to enhance the enrichment. Default is c(0, 1).}

\item{color}{Color for heatmap. Default is c("white", "coral1"), corresponding to the score from 0 to 1 or other ranges.}

\item{cell}{Cell width and height settings for heatmap. If set this parameter to a single number or \code{NA}, cell width and height will be the same, and \code{NA} means the cell will be adaptive to the input perturbation numbers. If set this parameter to a vector including two numbers or \code{NA}, the first one will be cell width and the second one will be height.}

\item{fontsize}{Fontsize of the heatmap. Default is 15.}

\item{angle}{Angle of the column labels, right now one can choose only from few predefined options (0, 45, 90, 270 and 315). Default is 90.}

\item{legend.title}{Include the title of legend in the heatmap or not. Default is FALSE.}

\item{plot.save}{Logical, save plots or not. Default is \code{TRUE}.}

\item{prefix}{Path to save the plots. Default is current directory.}

\item{label}{The prefix label of the output file. Notably, there needs a separator between default file names and the label, so label would be better to be like "label_". Default is "".}

\item{width}{Width of the graphics region of the pdf file in inches, for both png and pdf format. Default is 10.}

\item{height}{Height of the graphics region of the pdf file in inches, for both png and pdf format. Default is 8.}

\item{png_res}{The nominal resolution in ppi of png file. Higher png_res indicates a bigger and more clear png file. Default is 720.}
}
\description{
Calculate enrichment ratio for each perturbation in every cluster.
}
