% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_output.R
\name{config_generation}
\alias{config_generation}
\title{Generate HTML Config}
\usage{
config_generation(
  mtx,
  mtx_QC,
  sg_lib,
  score,
  pval,
  project,
  prefix = ".",
  label = "",
  species = "Hs",
  version = "v75",
  type = "ATAC",
  NTC = "NTC",
  article = "",
  data = "",
  article_name = "",
  data_name = "",
  gene_type = "Symbol",
  score_cut = 0.2,
  pval_cut = 0.05,
  DA = NULL,
  cicero = NULL,
  enhancer = NULL,
  base64 = FALSE
)
}
\arguments{
\item{mtx}{SeuratObject or directory to rds file of SeuratObject, with cell in columns and features in rows.}

\item{mtx_QC}{SeuratObject or directory to rds file after QC, with cell in columns and features in rows.}

\item{sg_lib}{Data frame or directory to a txt file containing 3 columns: cell, barcode, gene. If sgRNA information stored in a matrix-like format or input data frame only has sgRNA frequency of each cell, use \code{\link[SCREE]{sgRNAassign}} to assign sgRNA to each cell.}

\item{score}{Data frame or directory of score from \code{improved_scmageck_lr}.}

\item{pval}{Data frame or directory of p_value from \code{improved_scmageck_lr}.}

\item{project}{Project name.}

\item{prefix}{Path to the results generated by SCREE. Default is current directory.}

\item{label}{The prefix label of previous output file. Notably, there needs a separator between default file names and the label, so label would be better to be like "label_". Default is "".}

\item{species}{Only support "Hs" and "Mm". Default is "Hs".}

\item{version}{Version of the reference genome(Ensembl). Default is "v75".}

\item{type}{Data type, one of "RNA", "ATAC", "enhancer". Default is "ATAC".}

\item{NTC}{The name of negative controls. Default is "NTC".}

\item{article}{Which article the dataset from (URL).}

\item{data}{Data source.}

\item{article_name}{Name of the article.}

\item{data_name}{Name of the data.}

\item{gene_type}{Type of gene name, selected from one of c("Symbol", "Ensembl"). Default is "Symbol".}

\item{score_cut}{Score cutoff of \code{improved_scmageck_lr} results. Default is 0.2.}

\item{pval_cut}{P-value cutoff of \code{improved_scmageck_lr} results. Default is 0.05.}

\item{DA}{DA peak config generated by SCREE. Default is \code{NULL}.}

\item{cicero}{Cicero config generated by SCREE. Default is \code{NULL}.}

\item{enhancer}{Enhancer config generated by SCREE. Default is \code{NULL}.}
}
\description{
Generate config for HTML output, based on previous output of SCREE.
}
