% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_peaks.R
\name{DApeaks}
\alias{DApeaks}
\title{Find Differential Peaks}
\usage{
DApeaks(
  mtx,
  selected,
  NTC = "NTC",
  min.pct = 0.1,
  test.use = "wilcox",
  p_adj_cut = 0.05,
  logFC_cut = 0.25
)
}
\arguments{
\item{mtx}{SeuratObject or directory to rds file of SeuratObject, with cell in columns and features in rows.}

\item{selected}{Perturbation to find DA peaks.}

\item{NTC}{The name of negative controls. Default is "NTC".}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the selected perturbation or NTC. Meant to speed up the function by not testing genes that are very infrequently expressed. Default is 0.1.}

\item{test.use}{Denotes which test to use. Available options are: "wilcox", "bimod", "roc", "t", "negbinom", "poisson", "LR", "MAST", "DESeq2". See details from \code{\link[Seurat]{FindMarkers}}. Default is "wilcox".}

\item{p_adj_cut}{Maximum adjust p_value to . Default is 0.05.}

\item{logFC_cut}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.25. Increasing logfc.threshold speeds up the function, but can miss weaker signals.}
}
\description{
Find differential accessible peaks between selected perturbation and negative control. Based on \code{\link[Seurat]{FindMarkers}}.
}
