% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancer.R
\name{DirectTarget}
\alias{DirectTarget}
\title{Calculate Direct Target Numbers}
\usage{
DirectTarget(
  score,
  pval,
  selected = NULL,
  species = "Hs",
  version = "v75",
  gene_annotations = NULL,
  upstream = 2e+06,
  downstream = 2e+06,
  score_cut = 0.2,
  pval_cut = 0.05,
  table.save = TRUE,
  prefix = ".",
  label = ""
)
}
\arguments{
\item{score}{Data frame or directory of score from \code{improved_scmageck_lr}, genes in rows and perturbations in columns.}

\item{pval}{Data frame or directory of p_value from \code{improved_scmageck_lr}, genes in rows and perturbations in columns.}

\item{selected}{Enhancer regions to calculate direct target ratio. By default, all enhancers will be chosen.}

\item{species}{Only support "Hs" and "Mm". Default is "Hs".}

\item{version}{Version of the reference genome(Ensembl). Default is "v75".}

\item{gene_annotations}{Gene annotations stored in data frame format, including c("chromosome", "start", "end", "strand", "transcript") as colnames. Default is \code{NULL}, gene annotations are from \code{\link{ensembldb}}.}

\item{upstream}{The number of nucleotides upstream of the start site of selected region. Default is 2000000.}

\item{downstream}{The number of nucleotides downstream of the start site of selected region. Default is 2000000.}

\item{score_cut}{Score cutoff of \code{improved_scmageck_lr} results. Default is 0.2.}

\item{pval_cut}{P-value cutoff of \code{improved_scmageck_lr} results. Default is 0.05.}

\item{table.save}{Logical, save the results as table or not. Default is \code{TRUE}.}

\item{prefix}{Path to save the table. Default is current directory.}

\item{label}{The prefix label of the output file. Notably, there needs a separator between default file names and the label, so label would be better to be like "label_". Default is "".}
}
\description{
Identify potential direct targets for each enhancer.
}
