% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_peaks.R
\name{enhancer}
\alias{enhancer}
\title{Find Potential Enhancer List}
\usage{
enhancer(da_peak, gene_anno, overlap_cut, pro_up = 2000, pro_down = 0)
}
\arguments{
\item{da_peak}{Data frame of DA peaks list, generated from \code{\link[SCREE]{DApeaks}}.}

\item{overlap_cut}{Maximum overlap nucleotides between peaks and promoters. Default is 0.}

\item{pro_up}{The number of nucleotides upstream of the transcription start site that should be included in the promoter region. Default is 3000.}

\item{pro_down}{The number of nucleotides downstream of the transcription start site that should be included in the promoter region. Default is 0.}

\item{gene_annotations}{Gene annotations stored in data frame format, including c("chromosome", "start", "end", "strand", "transcript") as colnames. Default is \code{NULL}, gene annotations are from \code{\link{ensembldb}}.}
}
\description{
Identify potential enhancer list from DA peaks.
}
