% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancer.R
\name{EnhancerGeneExpression}
\alias{EnhancerGeneExpression}
\title{Gene Expression Plot}
\usage{
EnhancerGeneExpression(
  sg_lib,
  mtx,
  selected = NULL,
  species = "Hs",
  version = "v75",
  gene_annotations = NULL,
  upstream = 2e+06,
  downstream = 2e+06,
  NTC = "NTC",
  title.size = 25,
  x.text.size = 25,
  x.title.size = 22,
  y.text.size = 25,
  y.title.size = 22,
  annotate.size = 40,
  plot.save = TRUE,
  prefix = ".",
  width = 18,
  height = 20,
  png_res = 720,
  html_config = FALSE
)
}
\arguments{
\item{sg_lib}{Data frame or directory to a txt file containing 3 columns: cell, barcode, gene. If sgRNA information stored in a matrix-like format or input data frame only has sgRNA frequency of each cell, use \code{\link[SCREE]{sgRNAassign}} to assign sgRNA to each cell.}

\item{mtx}{SeuratObject or directory to rds file of SeuratObject, with cell in columns and features in rows.}

\item{selected}{Enhancer regions to visualize gene expression. By default, all enhancers will be chosen.}

\item{species}{Only support "Hs" and "Mm". Default is "Hs".}

\item{version}{Version of the reference genome(Ensembl). Default is "v75".}

\item{gene_annotations}{Gene annotations stored in data frame format, including c("chromosome", "start", "end", "strand", "transcript") as colnames. Default is \code{NULL}, gene annotations are from \code{\link{ensembldb}}.}

\item{upstream}{The number of nucleotides upstream of the start site of selected region. Default is 2000000.}

\item{downstream}{The number of nucleotides downstream of the start site of selected region. Default is 2000000.}

\item{NTC}{The name of negative controls. Default is "NTC".}

\item{title.size}{Numeric, title size of each gene expression plot. Default is 25.}

\item{x.text.size}{Numeric, x-axis text size of each gene expression plot. Default is 25.}

\item{x.title.size}{Numeric, x-axis title size of each gene expression plot. Default is 22.}

\item{y.text.size}{Numeric, y-axis text size of each gene expression plot. Default is 25.}

\item{y.title.size}{Numeric, y-axis title size of each gene expression plot. Default is 22.}

\item{annotate.size}{Numeric, title size of the whole plot including 12 gene expression plot. Default is 40.}

\item{plot.save}{Logical, save plots or not. Default is \code{TRUE}.}

\item{prefix}{Path to save the plots. Default is current directory.}

\item{width}{Width of the graphics region of the pdf file in inches, for both png and pdf format. Default is 18.}

\item{height}{Height of the graphics region of the pdf file in inches, for both png and pdf format. Default is 20.}

\item{png_res}{The nominal resolution in ppi of png file. Higher png_res indicates a bigger and more clear png file. Default is 720.}

\item{html_config}{Logical, generate and return a list includes the config character string of html. Default is \code{FALSE}.}
}
\description{
Plot expression of genes near each enhancer.
}
