% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame2indmatrix.R
\name{frame2indmatrix}
\alias{frame2indmatrix}
\title{function definitions ##### Create a 0-1 index matrix to store the infromation of
perturbations.0 means cells without the perturbation while 1 means cells with
the perturbation}
\usage{
frame2indmatrix(bc_d, targetobj)
}
\description{
function definitions ##### Create a 0-1 index matrix to store the infromation of
perturbations.0 means cells without the perturbation while 1 means cells with
the perturbation
}
