% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAC.R
\name{GetPromoter}
\alias{GetPromoter}
\title{Get Potential Promoter Regions}
\usage{
GetPromoter(
  species = "Hs",
  version = "v75",
  gene_type = "Symbol",
  protein_coding = TRUE,
  pro_up = 3000,
  pro_down = 0
)
}
\arguments{
\item{species}{Only support "Hs" and "Mm". Default is "Hs".}

\item{version}{Version of the reference genome(Ensembl). Default is "v75".}

\item{gene_type}{Type of gene name, selected from one of c("Symbol", "Ensembl"). Default is "Symbol".}

\item{protein_coding}{Logical, only use protein coding gene or not. Default is \code{TRUE}.}

\item{pro_up}{Numeric, the number of nucleotides upstream of the transcription start site that should be included in the promoter region. Default is 3000.}

\item{pro_down}{Numeric, the number of nucleotides downstream of the transcription start site that should be included in the promoter region. Default is 0.}
}
\description{
Get potential promoter regions for each gene, using specific reference.
}
