% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsolvedmatrix.R
\name{getsolvedmatrix}
\alias{getsolvedmatrix}
\title{function definitions ##### Return liner regression results of perturbation score}
\usage{
getsolvedmatrix(Xm, Ym, lambda = 0.01)
}
\description{
function definitions ##### Return liner regression results of perturbation score
}
