% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsolvedmatrix_with_permutation_cell_label.R
\name{getsolvedmatrix_with_permutation_cell_label}
\alias{getsolvedmatrix_with_permutation_cell_label}
\title{function definitions ##### Return list of 2 matrix:
liner regression results of perturbation score and
p-value calculated by permutation}
\usage{
getsolvedmatrix_with_permutation_cell_label(
  Xm,
  Ym,
  lambda = 0.01,
  npermutation = 1000
)
}
\description{
function definitions ##### Return list of 2 matrix:
liner regression results of perturbation score and
p-value calculated by permutation
}
