% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{Correlation Heatmap}
\usage{
heatmap(
  score,
  pval,
  score_cut = 0.2,
  pval_cut = 0.05,
  NTC = "NegCtrl",
  remove_neg = TRUE,
  num = 0,
  min.filtered.gene = 3,
  method = "pearson",
  color = c("#4DBBD5FF", "white", "#E64B35FF"),
  cell = "auto",
  fontsize = "auto",
  angle = 90,
  legend.title = FALSE,
  plot.save = TRUE,
  prefix = ".",
  label = "",
  width = "auto",
  height = "auto",
  png_res = 720
)
}
\arguments{
\item{score}{Data frame or directory of score from \code{improved_scmageck_lr}, genes in rows and perturbations in columns.}

\item{pval}{Data frame or directory of p_value from \code{improved_scmageck_lr}, genes in rows and perturbations in columns.}

\item{score_cut}{Score cutoff of \code{improved_scmageck_lr} results. Default is 0.2.}

\item{pval_cut}{P-value cutoff of \code{improved_scmageck_lr} results. Default is 0.05.}

\item{NTC}{The name of negative controls. Default is "NegCtrl".}

\item{remove_neg}{Logical, remove negative control before calculating correlation or not. Default is \code{TRUE}.}

\item{num}{Minimum number of perturbation passed score_cut or pval_cut for each gene. Only genes with at least \code{num} numbers of perturbation passed threshold will be retained for correlation calculation. Default is 0.}

\item{min.filtered.gene}{Minimum number of genes retained for correlation calculation. Default is 3.}

\item{method}{Method to calculate correlation, can be one of "spearman", "kendall" or "pearson". Default is "pearson".}

\item{color}{Color for heatmap. Default is c("#4DBBD5FF", "white", "#E64B35FF"), corresponding to the correlation from -1 to 1.}

\item{cell}{Cell width and height settings for heatmap. If set this parameter to a single number or \code{NA}, cell width and height will be the same, and \code{NA} means the cell will be adaptive to the input perturbation numbers. If set this parameter to a vector including two numbers or \code{NA}, the first one will be cell width and the second one will be height. Default is "auto", which means cell width and cell height will be set to 15 when the perturbations numbers > 20 (after removing negative control) while cell width and cell height will be set to NA when the perturbations numbers <= 20 (after removing negative control).}

\item{fontsize}{Fontsize of the heatmap. Default is "auto", which means the fontsize will be set to 15 when the perturbations numbers > 20 (after removing negative control) while the fontsize will be set to 10 when the perturbations numbers <= 20 (after removing negative control).}

\item{angle}{Angle of the column labels, right now one can choose only from few predefined options (0, 45, 90, 270 and 315). Default is 90.}

\item{legend.title}{Include the title of legend in the heatmap or not. Default is FALSE.}

\item{plot.save}{Logical, save plots or not. Default is \code{TRUE}.}

\item{prefix}{Path to save the plots. Default is current directory.}

\item{label}{The prefix label of the output file. Notably, there needs a separator between default file names and the label, so label would be better to be like "label_". Default is "".}

\item{width}{Width of the graphics region of the pdf file in inches, for both png and pdf format. Default is "atuo", according to the cellwidth unless cellwidth is \code{NA}. If cellwidth is \code{NA} or legend.title is \code{TRUE} and width is "auto", the width will be set to 10.}

\item{height}{Height of the graphics region of the pdf file in inches, for both png and pdf format. Default is "atuo", according to the cellheight unless cellheight is \code{NA}. If cellheight is \code{NA} or legend.title is \code{TRUE} and height is "auto", the height will be set to 8.}

\item{png_res}{The nominal resolution in ppi of png file. Higher png_res indicates a bigger and more clear png file. Default is 720.}
}
\description{
Calculate correlation between perturbations based on regulatory score of a union set of potential target genes for all perturbations. The plot function is based on \code{\link[pheatmap]{pheatmap}}
}
