% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_output.R
\name{html_output}
\alias{html_output}
\title{Generate HTML}
\usage{
html_output(html_dir, config, replace = 3, prefix = ".", label = "")
}
\arguments{
\item{html_dir}{Template HTML file directory.}

\item{config}{Config generated before.}

\item{replace}{Either the child number to return (by position), or the name of the child node to return. If there are multiple child nodes with the same name, the first will be returned}

\item{prefix}{Path to save the html. Default is current directory.}

\item{label}{The prefix label of the output file. Notably, there needs a separator between default file names and the label, so label would be better to be like "label_". Default is "".}
}
\description{
Generate HTML based on the config from \code{\link[SCREE]{config_generation}}.
}
