% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improved_scmageck_lr.R
\name{improved_scmageck_lr}
\alias{improved_scmageck_lr}
\title{function definitions ##### Improved scMAGeCK_LR, save and return two data.frame:
rownames indicates selected_gene while colnames indicates perturbations.}
\usage{
improved_scmageck_lr(
  BARCODE,
  RDS,
  NEGCTRL,
  SELECT_GENE = NULL,
  LABEL = NULL,
  PERMUTATION = NULL,
  SAVEPATH = "./",
  LAMBDA = 0.01,
  NTC_baseline = TRUE
)
}
\description{
function definitions ##### Improved scMAGeCK_LR, save and return two data.frame:
rownames indicates selected_gene while colnames indicates perturbations.
}
