% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCREEN.R
\name{SCREEN}
\alias{SCREEN}
\title{Batch Run}
\usage{
SCREEN(
  sg_dir,
  mtx_dir,
  fragments,
  cal.FRiP = TRUE,
  species = "Hs",
  version = "v75",
  data_type = "RNA",
  Mixscape = TRUE,
  prefix = "./",
  label = "",
  gene_type = "Symbol",
  protein_coding = TRUE,
  frac = 0.01,
  cal.mt = TRUE,
  nFeature = c(200, 5000),
  nCount = 1000,
  FRiP = 0.1,
  mt = 10,
  blank_NTC = FALSE,
  lambda = 0.01,
  permutation = NULL,
  p_val_cut = 0.05,
  score_cut = 0.5,
  cicero_p_val_cut = 0.05,
  cicero_score_cut = 0,
  ylimit = "auto",
  project = "perturb",
  NTC = "NTC",
  replicate = 1,
  select_gene = NULL,
  selected = NULL,
  gene_annotations = NULL,
  pro_annotations = NULL,
  pro_up = 3000,
  pro_down = 0,
  overlap_cut = 0,
  p_adj_cut = 0.05,
  logFC_cut = 1,
  min.pct = 0.2,
  upstream = 2e+06,
  downstream = 2e+06,
  test.use = "wilcox",
  track_size = c(1, 0.3, 0.2, 0.3),
  include_axis_track = TRUE,
  connection_color = "#7F7CAF",
  connection_color_legend = TRUE,
  connection_width = 2,
  connection_ymax = NULL,
  gene_model_color = "#81D2C7",
  alpha_by_coaccess = FALSE,
  gene_model_shape = c("smallArrow", "box")
)
}
\arguments{
\item{sg_dir}{Data frame or directory to a txt file containing 3 columns: cell, barcode, gene. If sgRNA information stored in a matrix-like format or sinput data frame only has sgRNA frequence of each cell, use \code{\link[SCREEN]{sgRNAassign}} to assign sgRNA to each cell.}

\item{mtx_dir}{SeuratObject or directory to rds file of SeuratObject, with cell in columns and features in rows.}

\item{fragments}{Directory of fragments file used to calculate FRiP for perturb-ATAC input.}

\item{cal.FRiP}{Logical, calculate FRiP or not. Default is \code{TRUE}.}

\item{species}{Only support "Hs" and "Mm", if input other species, \code{percent.mt} will be count as "Mm". Default is "Hs".}

\item{version}{Version of the reference genome(Ensembl), used for perturb-ATAC input and perturb-enhancer input. Default is "v75".}

\item{data_type}{Type of input data, can be one of c("RNA", "ATAC"). Default is "RNA".}

\item{Mixscape}{Logical, run \code{IntegratedMixscape} or not. Default is \code{TRUE}.}

\item{prefix}{Path to save all the results. Default is current directory.}

\item{label}{The label of the output file.}

\item{gene_type}{Type of gene name, selected from one of c("Symbol", "Ensembl"). Default is "Symbol".}

\item{protein_coding}{Logical, only use protein coding gene or not. This parameter is only used for calculating gene activity for perturb-ATAC input. Default is \code{TRUE}.}

\item{frac}{A paramter for filtering low expressed genes or low accessibility peaks. By default, only genes or peaks that have expressions or counts in at least that fractions of cells are kept. Default is 0.01.}

\item{cal.mt}{Logical, calculate percentage of mitochondrial gene expression of each cell or not. Default is \code{TRUE}.}

\item{nFeature}{Limitation of detected feature numbers in each cell, in the format c(200, 5000). Default is c(200, 5000).}

\item{nCount}{Minimal count numbers in each cell. Default is 1000.}

\item{FRiP}{Minimal FRiP of each cell. Default is 0.1.}

\item{mt}{Maximal percentage of mitochondrial gene expression of each cell. Default is 10.}

\item{blank_NTC}{Logical, use blank control as negative control or not. Default is \code{FALSE}.}

\item{lambda}{Parameter used in ridge regression of \code{improved_scmageck_lr}. Default is 0.01.}

\item{permutation}{Permutation times in \code{improved_scmageck_lr}. Default is 10000.}

\item{p_val_cut}{P-value cutoff of \code{improved_scmageck_lr} results. Default is 0.05.}

\item{score_cut}{Score cutoff of \code{improved_scmageck_lr} results. Default is 0.5.}

\item{cicero_p_val_cut}{P-value cutoff of \code{improved_scmageck_lr} results used for \code{ciceroPlot}. Default is 0.05.}

\item{ylimit}{Limitation of y-axis of \code{DE_gene_plot} in the format c(-600, 600, 200). These numbers mean c(minimum, maximum, interval). Default is "auto", which means that this function will get \code{ylimit} automatically.}

\item{project}{Title of \code{DE_gene_plot}. Default is "perturb".}

\item{NTC}{The name of the genes served as negative controls. Default is "NTC".}

\item{replicate}{Required a vector of replicate information corresponding to each cell with the same order. Default is 1, which means no replicate.}

\item{select_gene}{The list of genes for regression in \code{scMAGeCK} step. By default, all genes in the table are subject to regression.}

\item{selected}{Enhancer regions to visualize for perturb-enhancer or perturbations to chose for perturb-ATAC, in \code{cicero} step. By default, all enhancers or all perturbations will be chosen.}

\item{gene_annotations}{Gene annotations stored in data frame format, including c("chromosome", "start", "end", "strand", "transcript") as colnames, used for /code{ciceroPlot} step. By default, gene annotations are from \code{ensembldb}.}

\item{pro_annotations}{Gene annotations stored in data frame format, including c("chromosome", "start", "end", "strand", "transcript") as colnames. By default, gene annotations are from \code{ensembldb}.}

\item{pro_up}{The number of nucleotides upstream of the transcription start site that should be included in the promoter region, only used for perturb-ATAC data. Default is 3000.}

\item{pro_down}{The number of nucleotides downstream of the transcription start site that should be included in the promoter region, only used for perturb-ATAC data. Default is 0.}

\item{overlap_cut}{Maximum overlap nucleotides between peaks and promoters, only used for perturb-ATAC data. Default is 0.}

\item{p_adj_cut}{Parameter only used for finding DA peaks. Maximum adjust p_value calculated by \code{\link[Seurat]{FindMarkers}}. Default is 0.05.}

\item{logFC_cut}{Parameter only used for finding DA peaks. Minimum log fold change calculated by \code{\link[Seurat]{FindMarkers}}. Default is 1.}

\item{min.pct}{Parameter only used for finding DA peaks. Only test genes that are detected in a minimum fraction of min.pct cells in either of the NTC or perturbations. Meant to speed up the function by not testing genes that are very infrequently expressed. Default is 0.2.}

\item{upstream}{The number of nucleotides upstream of the start site of selected region in \code{ciceroPlot} step. Default is 2000000.}

\item{downstream}{The number of nucleotides downstream of the start site of selected region in \code{ciceroPlot} step. Default is 2000000.}

\item{test.use}{Parameter only used for finding DA peaks. Default is "wilcox".For more details, see \code{\link[Seurat]{FindMarkers}}.}

\item{track_size}{Size of each axis in /code{ciceroPlot} result. Default is c(1,.3,.2,.3). If \code{include_axis_track=FALSE}, track_size should be a vector with 3 elements.}

\item{include_axis_track}{Logical, should a genomic axis be plotted? Default is \code{TRUE}.}

\item{connection_color}{Color for connection lines. A single color, the name of a column containing color values, or the name of a column containing a character or factor to base connection colors on.}

\item{connection_color_legend}{Logical, should connection color legend be shown?}

\item{connection_width}{Width of connection lines.}

\item{connection_ymax}{Connection y-axis height. If NULL, chosen automatically.}

\item{gene_model_color}{Color for gene annotations.}

\item{alpha_by_coaccess}{Logical, should the transparency of connection lines be scaled based on co-accessibility score?}

\item{cicero_socre_cut}{Score cutoff of \code{improved_scmageck_lr} results used for \code{ciceroPlot}. Default is 0.}
}
\description{
Get all results using one function easily.
}
