% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{sgRNAassign}
\alias{sgRNAassign}
\title{Assign sgRNAs to Cells}
\usage{
sgRNAassign(
  sg_lib,
  type = "CountMatrix",
  row_names = FALSE,
  freq_cut = 20,
  freq_percent = 0.8,
  freq = "freq",
  cell = "cell",
  barcode = "barcode"
)
}
\arguments{
\item{sg_lib}{Data frame or count matrix or directory to a txt file containing information of sgRNA count in each cell.}

\item{type}{Type of input data frame, can be one of "CountMatrix" or "DataFrame". "CountMatrix" means a sgRNA count matrix similar to gene count matrix, whose rows are sgRNAs and columns are cells. "DataFrame" means a data frame includes three columns: "freq", "cell", "barcode". "freq" is sgRNA count in each cell, "barcode" is sgRNA name. Default is \code{"CountMatrix"}.}

\item{row_names}{Logical, is rownames included in the txt file or not. Default is \code{FALSE}.}

\item{freq_cut}{Cutoff of sgRNA count numbers. For each cell, only sgRNA with counts more than freq_cut will be retained. Default is \code{20}}

\item{freq_percent}{Cutoff of sgRNA frequency. For each cell, only sgRNA with frequency more than freq_percent will be retained. Default is \code{0.8}.}

\item{freq}{Name of the "freq" column, only used for "DataFrame" type. Default is \code{"freq"}.}

\item{cell}{Name of the "cell" column, only used for "DataFrame" type. Default is \code{"cell"}.}

\item{barcode}{Name of the "barcode" column, only used for "DataFrame" type. Default is \code{"barcode"}.}
}
\description{
Assign sgRNAs to cells based on sgRNA count matrix or data frame includes sgRNA counts in each cell.
}
