% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{Volcano Plot For Potential DEGs}
\usage{
volcano(
  score,
  pval,
  selected = NULL,
  pval_cut = 0.05,
  score_cut = 0.2,
  color = c("#4DBBD5FF", "grey", "#E64B35FF"),
  showCategory = 10,
  title.size = 23,
  legend.text.size = 16,
  legend.title.size = 20,
  x.text.size = 16,
  x.title.size = 20,
  y.text.size = 16,
  y.title.size = 20,
  pt.size = 2,
  raster = FALSE,
  plot.save = TRUE,
  prefix = ".",
  label = "",
  width = 7,
  height = 7,
  png_res = 720
)
}
\arguments{
\item{score}{Data frame or directory of score from \code{improved_scmageck_lr}, genes in rows and perturbations in columns.}

\item{pval}{Data frame or directory of p_value from \code{improved_scmageck_lr}, genes in rows and perturbations in columns.}

\item{selected}{Perturbation to visualize. Default is \code{NULL}, all perturbations will be chosen.}

\item{pval_cut}{P-value cutoff of \code{improved_scmageck_lr} results. Default is 0.05.}

\item{score_cut}{Score cutoff of \code{improved_scmageck_lr} results. Default is 0.2.}

\item{color}{Color for volcano plot. Default is c("#4DBBD5FF", "grey", "#E64B35FF"), corresponding to "down-regulated", "no sense" and "up-regulated".}

\item{showCategory}{Numbers of genes to be labeled for both "down" and "up". Default is 10, and will label 10 "up-regulated" genes and 10 "down-regulated" genes.}

\item{title.size}{Numeric, title size of the volcano plot. Default is 23.}

\item{legend.text.size}{Numeric, legend text size of the volcano plot. Default is 16.}

\item{legend.title.size}{Numeric, legend title size of the volcano plot. Default is 20.}

\item{x.text.size}{Numeric, x-axis text size of the volcano plot. Default is 16.}

\item{x.title.size}{Numeric, x-axis title size of the volcano plot. Default is 20.}

\item{y.text.size}{Numeric, y-axis text size of the volcano plot. Default is 16.}

\item{y.title.size}{Numeric, y-axis title size of the volcano plot. Default is 20.}

\item{pt.size}{Point size of the volcano plot. Default is 2.}

\item{raster}{Logical, convert points to raster format, will be useful to reduce the storage cost of the output figure or pdf. Default is \code{FALSE}.}

\item{plot.save}{Logical, save plots or not. Default is \code{TRUE}.}

\item{prefix}{Path to save the plots. Default is current directory.}

\item{label}{The prefix label of the output file. Notably, there needs a separator between default file names and the label, so label would be better to be like "label_". Default is "".}

\item{width}{Width of the graphics region of the pdf file in inches, for both png and pdf format. Default is 7.}

\item{height}{Height of the graphics region of the pdf file in inches, for both png and pdf format. Default is 7.}

\item{png_res}{The nominal resolution in ppi of png file. Higher png_res indicates a bigger and more clear png file. Default is 720.}
}
\description{
Visualize regulatory score and p-value distribution for each perturbation.
}
