% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegratedMixscape.R
\name{IntegratedMixscape}
\alias{IntegratedMixscape}
\title{function definitions ##### Integrated some functions in Seurat as mentioned in
vignette of Mixscape.Prefer Seurat object after QC.}
\usage{
IntegratedMixscape(
  mtx_dir,
  sg_dir,
  sg.to.use = "all",
  perturb.sig.only = FALSE,
  mixscape.only = FALSE,
  nfeature = 2000,
  dims = 1:40,
  assays = "RNA",
  algorithm = 1,
  label.cut = 20,
  resolution = 0.8,
  NTC = "NTC",
  prefix = "./",
  label = "",
  plot.save = TRUE,
  plot.show = TRUE
)
}
\description{
function definitions ##### Integrated some functions in Seurat as mentioned in
vignette of Mixscape.Prefer Seurat object after QC.
}
