% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_scale.R
\name{normalize_scale}
\alias{normalize_scale}
\title{function definitions ##### Normalize and scale data.}
\usage{
normalize_scale(mtx_dir)
}
\description{
function definitions ##### Normalize and scale data.
}
