% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgRNA_quality_plot.R
\name{sgRNA_quality_plot}
\alias{sgRNA_quality_plot}
\title{function definitions ##### Plot for sgRNA information.}
\usage{
sgRNA_quality_plot(sg_dir, mtx_dir, LABEL = "", prefix = "./")
}
\description{
function definitions ##### Plot for sgRNA information.
}
